requirements.txt
# Function dependencies, for example:
# package>=version
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
google-api-core==1.30.0
google-api-python-client==2.8.0
google-auth==1.30.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
google-cloud-storage
glom==20.11.0
httplib2==0.19.1
idna==2.10
numpy==1.20.3
packaging==20.9
pandas==1.2.4
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyMySQL==1.0.2
requests==2.25.1
rsa==4.7.2
six==1.16.0
uritemplate==3.0.1
urllib3==1.26.5
sqlalchemy
gunicorn==20.1.0
pg8000==1.22.0
pyodbc==4.0.32
sqlalchemy-pytds==0.3.3
